fix: create utils to translate aria-label for select input and widget
